Replace rdma_read_min_ack_part with rdma_read_min_ack
The number of rdma read requests that can be posted (per connection) appears to
be on the order of 16, which can easily result in acknowledging one completion
at a time for relatively small values of rdma_read_min_ack_part. Replacing the
configuration parameter with a simple number of requests to acknowledge will
likely result in fewer of this sort of bad configuration.